Creating a method by which the velocity of the boids is limited, so it doesn't go out of control.
Fixed the issue which was limiting the number of possible boids. The limit is now approx. 60 boids.

The issue was that the positions were being allocated a position on a screen with a division by rand()%10.....As the number of boids grew, there was an increasing chance that this calcualtion woujld include a division-by-zero error.