Boid-avoidance algorithm has been fully implemented, and also tuned. Centre-seeking algorithm has been fully implemented, and somewhat tuned.
A complicated equation ahs been implemented which allows for casual steering of the boids, and yet allows for more extreme avoidance turning in close-call situations.

The ability to increment the centripetal and avoidance turning forces has been implemented (WASD/arrows).

A word document has been included which documents the design decisions of some of the working parts of this program.